+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 6.2.0 (unreleased)
 
+### Breaking changes
+
+* **Gauge Panel**: The suffix / prefix options have been removed from the new Guage Panel (introduced in v6.0). [#16870](https://github.com/grafana/grafana/issues/16870). 
+
 # 6.1.6 (2019-04-29)
 ### Features / Enhancements
 * **Security**: Bump jQuery to 3.4.0 . [#16761](https://github.com/grafana/grafana/pull/16761), [@dprokop](https://github.com/dprokop)